Fix root cause of iframe timeout
