+++ src/test/playground-ide_test.ts
@@ -8,19 +8,19 @@ import {assert} from '@esm-bundle/chai';
 import {html, render} from 'lit';
 import {PlaygroundIde} from '../playground-ide.js';
 import '../playground-ide.js';
-import {sendKeys} from '@web/test-runner-commands';
+import {sendKeys, executeServerCommand} from '@web/test-runner-commands';
 
 import type {ReactiveElement} from '@lit/reactive-element';
 import type {PlaygroundCodeEditor} from '../playground-code-editor.js';
 import type {PlaygroundProject} from '../playground-project.js';
 import type {PlaygroundFileEditor} from '../playground-file-editor.js';
+import type {PlaygroundPreview} from '../playground-preview.js';
 
 // There is browser variability with zero width spaces. This helper keeps tests
 // consistent.
 function innerTextWithoutSpaces(el?: HTMLElement | null): string {
   return el?.innerText.replace(/[\u200B\s]/g, '') ?? '';
 }
-
 suite('playground-ide', () => {
   let container: HTMLDivElement;
   let testRunning: boolean;
@@ -602,6 +602,98 @@ suite('playground-ide', () => {
     assert.equal(tokenUnderCursor.start, 0);
     assert.equal(tokenUnderCursor.end, 7);
     assert.equal(tokenUnderCursor.string, 'console');
+  });
+
+  test('reloading preview does not modify history', async () => {
+    const historyLengthBefore = window.history.length;
+
+    // NOTE: For some reason, the parent window's history only seems to be
+    // affected when the iframe origin is different.
+    const separateOrigin = (await executeServerCommand(
+      'separate-origin'
+    )) as string;
+
+    render(
+      html`
+        <playground-ide sandbox-base-url="${separateOrigin}">
+          <script type="sample/html" filename="index.html">
+            <body>
+              <p>Hello HTML 1</p>
+            </body>
+          </script>
+        </playground-ide>
+      `,
+      container
+    );
+    const iframe = (await pierce(
+      'playground-ide',
+      'playground-preview',
+      'iframe'
+    )) as HTMLIFrameElement;
+    await waitForIframeLoad(iframe);
+
+    const editor = (await pierce(
+      'playground-ide',
+      'playground-file-editor',
+      'playground-code-editor'
+    )) as PlaygroundCodeEditor;
+    updateCurrentFile(editor, 'Hello HTML 2');
+
+    const project = (await pierce(
+      'playground-ide',
+      'playground-project'
+    )) as PlaygroundProject;
+    project.save();
+    await waitForIframeLoad(iframe);
+
+    const historyLengthAfter = window.history.length;
+    assert.equal(historyLengthAfter, historyLengthBefore);
+  });
+
+  test('reloading preview does not create a new iframe element', async () => {
+    render(
+      html`
+        <playground-ide>
+          <script type="sample/html" filename="index.html">
+            <body>
+              <p>Hello HTML 1</p>
+            </body>
+          </script>
+        </playground-ide>
+      `,
+      container
+    );
+    const preview = (await pierce(
+      'playground-ide',
+      'playground-preview'
+    )) as PlaygroundPreview;
+
+    const iframe = preview.iframe!;
+
+    await waitForIframeLoad(iframe);
+
+    const editor = (await pierce(
+      'playground-ide',
+      'playground-file-editor',
+      'playground-code-editor'
+    )) as PlaygroundCodeEditor;
+    updateCurrentFile(editor, 'Hello HTML 2');
+
+    const project = (await pierce(
+      'playground-ide',
+      'playground-project'
+    )) as PlaygroundProject;
+
+    await project.save();
+    await waitForIframeLoad(iframe);
+
+    const newIframe = (await pierce(
+      'playground-ide',
+      'playground-preview',
+      'iframe'
+    )) as HTMLIFrameElement;
+
+    assert.equal(newIframe, iframe);
   });
 
   test('delete file using menu', async () => {
